requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic_settings
mysql-connector-python
Pygment
pymysql